Remove containers/Diagram.ts
