--- src/bp/ui-studio/src/web/views/FlowBuilder/containers/Diagram.ts
@@ -1,58 +1,0 @@
-import { connect } from 'react-redux'
-import {
-  buildNewSkill,
-  closeFlowNodeProps,
-  copyFlowNode,
-  createFlow,
-  createFlowNode,
-  fetchFlows,
-  insertNewSkillNode,
-  openFlowNodeProps,
-  pasteFlowNode,
-  removeFlowNode,
-  setDiagramAction,
-  switchFlow,
-  switchFlowNode,
-  updateFlow,
-  updateFlowNode,
-  updateFlowProblems
-} from '~/actions'
-import { getCurrentFlow, getCurrentFlowNode } from '~/reducers'
-
-import Diagram from '../diagram'
-
-const mapStateToProps = state => ({
-  flows: state.flows,
-  currentFlow: getCurrentFlow(state),
-  currentFlowNode: getCurrentFlowNode(state),
-  currentDiagramAction: state.flows.currentDiagramAction,
-  canPasteNode: Boolean(state.flows.nodeInBuffer),
-  skills: state.skills.installed
-})
-
-const mapDispatchToProps = {
-  fetchFlows,
-  switchFlowNode,
-  openFlowNodeProps,
-  closeFlowNodeProps,
-  setDiagramAction,
-  createFlowNode,
-  removeFlowNode,
-  createFlow,
-  updateFlowNode,
-  switchFlow,
-  updateFlow,
-  copyFlowNode,
-  pasteFlowNode,
-  insertNewSkillNode,
-  updateFlowProblems,
-  buildSkill: buildNewSkill
-}
-
-const ConnectedDiagram = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-  null,
-  { withRef: true }
-)(Diagram)
-export default ConnectedDiagram
